# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "quant-agent-backend",
    "description": "app API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/memory/memories": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Create Memory",
        "description": "Create a new user memory.",
        "operationId": "create_memory_api_memory_memories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMemoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/memory/facts": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Create Fact",
        "description": "Create a new memory fact.",
        "operationId": "create_fact_api_memory_facts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryFactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/memory/facts/{fact_id}": {
      "delete": {
        "tags": [
          "memory"
        ],
        "summary": "Delete Fact",
        "description": "Delete a memory fact.",
        "operationId": "delete_fact_api_memory_facts__fact_id__delete",
        "parameters": [
          {
            "name": "fact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Fact Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FactCreateRequest": {
        "properties": {
          "fact_type": {
            "type": "string",
            "title": "Fact Type",
            "description": "Type of fact (preference, knowledge, plan, relationship)"
          },
          "content": {
            "type": "string",
            "maxLength": 4096,
            "title": "Content",
            "description": "Fact content"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding",
            "description": "Vector embedding for similarity"
          }
        },
        "type": "object",
        "required": [
          "fact_type",
          "content"
        ],
        "title": "FactCreateRequest",
        "description": "Request to create a memory fact."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryCreateRequest": {
        "properties": {
          "memory_type": {
            "type": "string",
            "title": "Memory Type",
            "description": "Type of memory (preference, fact, context, goal)"
          },
          "content": {
            "type": "string",
            "maxLength": 4096,
            "title": "Content",
            "description": "Memory content"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Confidence score 0-1",
            "default": 1.0
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Source of memory (chat, profile, explicit)"
          }
        },
        "type": "object",
        "required": [
          "memory_type",
          "content"
        ],
        "title": "MemoryCreateRequest",
        "description": "Request to create a user memory."
      },
      "MemoryFactOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "fact_type": {
            "type": "string",
            "title": "Fact Type"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "fact_type",
          "content",
          "embedding",
          "created_at",
          "updated_at"
        ],
        "title": "MemoryFactOut",
        "description": "Response DTO for a memory fact record."
      },
...
      "UserMemoryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "memory_type": {
            "type": "string",
            "title": "Memory Type"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "memory_type",
          "content",
          "confidence",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "UserMemoryOut",
        "description": "Response DTO for a user memory record."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "JWT issued by /api/v1/auth/login, set as the access_token cookie. Configure the value in Postman's Cookies modal for the host above."
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local dev server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ]
}